Move all dialects to separate directory
