+++ src/languages/bigquery/bigquery.formatter.ts
