+++ src/languages/redshift/redshift.formatter.ts
