+++ src/sqlFormatter.ts
@@ -1,16 +1,16 @@
-import BigQueryFormatter from 'src/languages/bigquery.formatter';
-import Db2Formatter from 'src/languages/db2.formatter';
-import HiveFormatter from 'src/languages/hive.formatter';
-import MariaDbFormatter from 'src/languages/mariadb.formatter';
-import MySqlFormatter from 'src/languages/mysql.formatter';
-import N1qlFormatter from 'src/languages/n1ql.formatter';
-import PlSqlFormatter from 'src/languages/plsql.formatter';
-import PostgreSqlFormatter from 'src/languages/postgresql.formatter';
-import RedshiftFormatter from 'src/languages/redshift.formatter';
-import SparkFormatter from 'src/languages/spark.formatter';
-import SqliteFormatter from 'src/languages/sqlite.formatter';
-import SqlFormatter from 'src/languages/sql.formatter';
-import TSqlFormatter from 'src/languages/tsql.formatter';
+import BigQueryFormatter from 'src/languages/bigquery/bigquery.formatter';
+import Db2Formatter from 'src/languages/db2/db2.formatter';
+import HiveFormatter from 'src/languages/hive/hive.formatter';
+import MariaDbFormatter from 'src/languages/mariadb/mariadb.formatter';
+import MySqlFormatter from 'src/languages/mysql/mysql.formatter';
+import N1qlFormatter from 'src/languages/n1ql/n1ql.formatter';
+import PlSqlFormatter from 'src/languages/plsql/plsql.formatter';
+import PostgreSqlFormatter from 'src/languages/postgresql/postgresql.formatter';
+import RedshiftFormatter from 'src/languages/redshift/redshift.formatter';
+import SparkFormatter from 'src/languages/spark/spark.formatter';
+import SqliteFormatter from 'src/languages/sqlite/sqlite.formatter';
+import SqlFormatter from 'src/languages/sql/sql.formatter';
+import TSqlFormatter from 'src/languages/tsql/tsql.formatter';
 
 import type { FormatOptions } from './types';
 import { ParamItems } from './formatter/Params';
