+++ test/db2.test.ts
@@ -1,7 +1,7 @@
 import dedent from 'dedent-js';
 
 import { format as originalFormat, FormatFn } from 'src/sqlFormatter';
-import Db2Formatter from 'src/languages/db2.formatter';
+import Db2Formatter from 'src/languages/db2/db2.formatter';
 import behavesLikeSqlFormatter from './behavesLikeSqlFormatter';
 
 import supportsAlterTable from './features/alterTable';
